X3: Remove redundant `parse_into_container` specialization
The `sequence` specialization is redundant, it actually hurts performance,
because it creates on every kleene/list/plus loop a new container, parses
a single value into it and then appends the content of this new container
to the actual container attribute.

This also fixes problems with handling single element sequences for
sequence parser.